refactor(tests): consolidate fragmented test suites, hoist inline imports, and eliminate duplicate tests
- Consolidate topological joint sorting tests from test_graph_topological.py into test_graph.py
- Merge root link and cyclic structure tests from test_validation.py into test_robot.py
- Disaggregate test_utils.py into test_name_sync_handler.py, test_mesh_io.py, and new test_decorators.py for clean 1-to-1 module mapping
- Eliminate duplicate tests in test_link_ops.py, test_blender_to_core.py, and test_validation_edge_cases.py
- Hoist safe inline imports across 17 test files to top-level module scope
- Standardize banner comments to 78-char borders in constants.py and update docs
