+++ docs/source/reference/api/blender.md
@@ -38,6 +38,11 @@ User actions and commands.
    :members:
    :undoc-members:
    :show-inheritance:
+
+.. automodule:: linkforge.blender.operators.selection_ops
+   :members:
+   :undoc-members:
+   :show-inheritance:
 ```
 
 ## Panels
@@ -46,11 +51,6 @@ UI Panels for managing robot data.
 
 ```{eval-rst}
 .. automodule:: linkforge.blender.panels.forge_panel
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. automodule:: linkforge.blender.panels.robot_panel
    :members:
    :undoc-members:
    :show-inheritance:
@@ -86,6 +86,11 @@ UI Panels for managing robot data.
 Blender scene properties for storing robot data.
 
 ```{eval-rst}
+.. automodule:: linkforge.blender.properties.geom_props
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 .. automodule:: linkforge.blender.properties.robot_props
    :members:
    :undoc-members:
@@ -132,7 +137,28 @@ Conversion between Blender and core models.
    :undoc-members:
    :show-inheritance:
 
+.. automodule:: linkforge.blender.adapters.geometry_extractor
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 .. automodule:: linkforge.blender.adapters.mesh_io
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## Logic
+
+Internal domain logic for asynchronous model building and collision synthesis.
+
+```{eval-rst}
+.. automodule:: linkforge.blender.logic.asynchronous_builder
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+.. automodule:: linkforge.blender.logic.collision_builder
    :members:
    :undoc-members:
    :show-inheritance:
@@ -143,6 +169,11 @@ Conversion between Blender and core models.
 Blender-specific helpers.
 
 ```{eval-rst}
+.. automodule:: linkforge.blender.utils.mode_guard
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 .. automodule:: linkforge.blender.utils.property_helpers
    :members:
    :undoc-members:
