+++ tests/unit/core/test_robot.py
@@ -927,3 +927,42 @@ class TestRobot:
                 _ = robot.root_link
             assert exc.value.code == ValidationErrorCode.NOT_FOUND
             assert "exists in graph but not in link index" in str(exc.value)
+
+
+class TestRobotRootAndCycles:
+    """Tests for Robot root link validation and cycle detection."""
+
+    def test_tree_structure_validation(self) -> None:
+        """Test that disconnected links are detected."""
+        robot = Robot(name="disconnected")
+        robot.add_link(Link(name="base"))
+        robot.add_link(Link(name="island"))
+
+        with pytest.raises(RobotValidationError) as exc:
+            robot.get_root_link()
+        assert exc.value.code == ValidationErrorCode.MULTIPLE_ROOTS
+
+    def test_cyclic_dependency_validation(self) -> None:
+        """Test that cycles in joint graph are detected."""
+        robot = Robot(name="cyclic")
+        robot.add_link(Link(name="l1"))
+        robot.add_link(Link(name="l2"))
+
+        robot.add_joint(Joint(name="j1", type=JointType.FIXED, parent="l1", child="l2"))
+        robot.add_joint(Joint(name="j2", type=JointType.FIXED, parent="l2", child="l1"))
+
+        assert robot.has_cycle is True
+
+        with pytest.raises(RobotValidationError) as exc:
+            robot.get_root_link()
+        assert exc.value.code == ValidationErrorCode.NO_ROOT
+
+    def test_multiple_root_links_validation(self) -> None:
+        """Test that multiple links without parents (roots) are detected."""
+        robot = Robot(name="multi_root")
+        robot.add_link(Link(name="root1"))
+        robot.add_link(Link(name="root2"))
+
+        with pytest.raises(RobotValidationError) as exc:
+            robot.get_root_link()
+        assert exc.value.code == ValidationErrorCode.MULTIPLE_ROOTS
